Added code and added a file.
Added some commands which I'm going to change later with a command handler but I didn't have time for that today.

Added all the code form google on  how to access a google sheet with node. Didn't do anything there myself.

Is making index.js run a function for every message a good thing? I hop so, because that's what I did there.

Last I added a file for all the data from the sheet to go into.